requirements.txt
# Python 3.6
Django==1.11
Pillow==5.3.0
PyMySQL==0.9.3
pytz==2018.7
alipay-sdk-python==3.2.53